+++ extensions/pipe-commands/assets/commands/search/github-search.sh
@@ -2,7 +2,8 @@
 
 # @raycast.title Github Search
 # @raycast.packageName Web Searches
-# @raycast.icon Globe
+# @raycast.icon ./icons/github.png
+# @raycast.iconDark ./icons/github-dark.png
 # @raycast.mode silent
 # @raycast.argument1 {"type": "text", "percentEncoded": true}