+++ extensions/pipe-commands/assets/commands/search/google-search.sh
@@ -2,7 +2,7 @@
 
 # @raycast.title Google Search
 # @raycast.packageName Web Searches
-# @raycast.icon Globe
+# @raycast.icon https://upload.wikimedia.org/wikipedia/commons/5/53/Google_%22G%22_Logo.svg
 # @raycast.mode silent
 # @raycast.argument1 {"type": "text", "percentEncoded": true}