+++ extensions/pipe-commands/assets/commands/search/wikipedia-search.sh
@@ -2,7 +2,7 @@
 
 # @raycast.title Wikipedia Search
 # @raycast.packageName Web Searches
-# @raycast.icon Globe
+# @raycast.icon https://upload.wikimedia.org/wikipedia/commons/8/80/Wikipedia-logo-v2.svg
 # @raycast.mode silent
 # @raycast.argument1 {"type": "text", "percentEncoded": true}