+++ extensions/pipe-commands/assets/commands/text/accessibility/read.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+# @raycast.title Read Text
+# @raycast.packageName Text Actions
+# @raycast.icon 💬
+# @raycast.argument1 {"type": "text"}
+# @raycast.mode silent
+
+say "$1"