+++ extensions/pipe-commands/assets/commands/url/extract-domain-from-url.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 # @raycast.title Extract Domain from URL
-# @raycast.icon Link
+# @raycast.icon 🔗
 # @raycast.packageName Developer Utils
 # @raycast.mode copy
 # @raycast.argument1 {"type": "text"}