+++ extensions/pipe-commands/assets/commands/url/urlencode.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # @raycast.title URL Encode
-# @raycast.icon Link
+# @raycast.icon 🔗
 # @raycast.packageName Developer Utils
 # @raycast.mode replace
 # @raycast.argument1 {"type": "text", "percentEncoded": true}